requirements.txt
fastapi
uvicorn
requests
pydantic
python-multipart
pillow
opencv-python
imagekitio
runpod